+++ scripts/archiso-extra-211204.sh
@@ -1,3 +1,17 @@
+mkdir -p airootfs/etc/pacman.d/hooks
+cat > airootfs/etc/pacman.d/hooks/vivaldi-install.hook << /cat
+# remove from airootfs!
+[Trigger]
+Operation = Install
+Type = Package
+Target = vivaldi
+
+[Action]
+Description = Set default web browser to Vivaldi...
+When = PostTransaction
+Exec = xdg-settings set default-web-browser vivaldi-stable.desktop
+/cat
+
 cat >> packages.x86_64 << /cat
 base-devel
 code