+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 - update `archiso-base-211205.sh`
 - update `archiso-fixes-211208.sh`
+- set default web browser to Vivaldi
 
 ### Fixed
 - fix hook in archiso fixes script